LayoutTests/fast/dom/HTMLDataGridElement/DataGridColumns-dom-expected.txt

CONSOLE MESSAGE: line 26: TypeError: Result of expression 'datagrid.columns' [undefined] is not an object.
FAIL: Timed out waiting for notifyDone to be called
FAIL: Initial value of datagrid.columns was null.
